+++ src/constants/denom.ts
@@ -4,6 +4,7 @@ export const DENOM_TO_DISPLAY_MAP: { [key in SupportedDenoms]: string } = {
   nhash: 'Nano Hash',
   hash: 'Hash',
   exchangesc: 'exchangesc',
+  cfigureomni: 'cfigureomni',
   usd: 'USD',
 };