+++ src/types/denoms.ts
@@ -1,1 +1,1 @@
-export type SupportedDenoms = 'hash' | 'nhash' | 'exchangesc' | 'usd';
+export type SupportedDenoms = 'hash' | 'nhash' | 'exchangesc' | 'usd' | 'cfigureomni';