+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Support for `cfigureomni` denom
+
+### Fixed
+
+- Eslint config updated per https://github.com/prettier/eslint-config-prettier/blob/main/CHANGELOG.md#version-800-2021-02-21
+
 ## 0.4.0
 
 ### Added